% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deetree.R
\name{deetree}
\alias{deetree}
\title{deetree}
\usage{
deetree()
}
\value{
A Json Tree
}
\description{
This function function plots the DEEDados tree. It is useful for exploring the available datasets at DEEDados.
}
\examples{
deetree()

}
