% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoagregall.R
\name{geoagregall}
\alias{geoagregall}
\title{Integrated Geographic Aggregation Composition}
\usage{
geoagregall(select_year = NULL)
}
\arguments{
\item{select_year}{Numeric (optional).}
}
\value{
tibble
}
\description{
Geographic aggregations for all municipalites in Rio Grande do Sul.
}
\examples{
geoagregall()
geoagregall(select_year = 2022)

}
