% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoagregall2.R
\name{geoagregall2}
\alias{geoagregall2}
\title{Integrated Geographic Aggregation Composition}
\usage{
geoagregall2(period = NULL)
}
\arguments{
\item{period}{Numeric (optional).}
}
\value{
tibble
}
\description{
Geographic aggregations for all municipalites in Rio Grande do Sul.
}
\examples{
geoagregall2()
geoagregall2(period = 2022)

}
