% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoagregcomp.R
\name{geoagregcomp}
\alias{geoagregcomp}
\title{Geographic Aggregation Composition}
\usage{
geoagregcomp(ag, id, period)
}
\arguments{
\item{ag}{Aggregation category. There are  three valid options:
\itemize{
\item "corede": for coredes, a state-specific planning regionalization.
\item "meso": for IBGE's mesoregions.
\item "micro": for IBGE's microregions.
}}

\item{id}{Greographic unit ID}

\item{period}{Year}
}
\value{
data.frame
}
\description{
The geographic aggregation composition. Be aware that municipalities do not have
}
\examples{
geoagregcomp(ag = "meso", id = 10, period = 2020)

}
