% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vardetail.R
\name{vardetails}
\alias{vardetails}
\title{Check vars details}
\usage{
vardetails(id)
}
\arguments{
\item{id}{}
}
\value{
A list
}
\description{
Check the details of a variable, by searching for it's id.
}
\examples{
vardetails(id = 4049)

}
