% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoagreg.R
\name{geoagreg}
\alias{geoagreg}
\title{Geographic Aggregation}
\usage{
geoagreg(ag = "municipio", sort = "ASC")
}
\arguments{
\item{ag}{Character. The aggregation category. There are  five valid options:
\itemize{
\item "municipio" (the default): for municipalities.
\item "corede": for coredes, a state-specific planning regionalization.
\item "meso": for IBGE's mesoregions.
\item "micro": for IBGE's microregions.
\item "estado": for the state of Rio Grande do Sul.
}}

\item{sort}{Character. This is used for sorting names from A to Z ("ASC") or Z to A ("DESC").Default is "ASC".}
}
\value{
a data.frame.
}
\description{
Check ID and name of geographic aggregations. For municipalities, latitude and longitude are also available.
}
\examples{
geoagreg()
geoagreg(ag = "municipio")
geoagreg(ag = "corede", sort = "DESC")

}
