% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoagregcomp.R
\name{geoagregcomp}
\alias{geoagregcomp}
\title{Geographic Aggregation Composition}
\usage{
geoagregcomp(ag = "corede", geo_id, period = 2022)
}
\arguments{
\item{ag}{Character. The aggregation category. There are  three valid options:
\itemize{
\item "corede": for coredes (the default), a state-specific planning regionalization only applied to Rio Grande do Sul.
\item "meso": for IBGE's mesoregions.
\item "micro": for IBGE's microregions.
}}

\item{geo_id}{Numeric. The Greographic unit ID}

\item{period}{Numeric. The year of the aggregation composition. This is needed only because coredes' composition changed over time. Default is 2022.}
}
\value{
data.frame
}
\description{
The geographic aggregation composition. Because municipalities are the minimum aggregation level available
}
\examples{
#First run geoagreg() to check the aggregation id of interest
geoagreg(ag = "meso")
#Let's get id=1 (Centro Ocidental Rio-Grandense) and year 2020 as example.
#Run:
geoagregcomp(ag = "meso", geo_id = 10, period = 2020)

}
