% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{getdata}
\alias{getdata}
\title{Download data from DEEDados}
\usage{
getdata(id = id, ag = ag, period = periodo, sort = "ASC")
}
\arguments{
\item{id}{The variable's ID.}

\item{ag}{The regional aggregation. One of "municipio", "corede", "microrregiao", "mesorregiao".}

\item{period}{The year to consult.}

\item{sort}{If the user wants to sort from "ASC" for ascendent order or "DESC" for descendent order. Default is "ASC".}
}
\value{
a data.frame
}
\description{
This function allows the user to download data from DEEDados API.
}
\examples{
my_data <- getdata(id = 4845,ag = "corede",period = c(2016,2017))
print(my_data)
print(my_data$data)

}
