% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vardetails.R
\name{vardetails}
\alias{vardetails}
\title{Check vars details}
\usage{
vardetails(var_id)
}
\arguments{
\item{var_id}{Variable's ID. It can be a single string or a vector.}
}
\value{
A tibble
}
\description{
Check the details of a variable, by searching for it's id.
}
\examples{
vardetails(var_id = 4049)
vardetails(var_id = c(4049,3755, 4784))

}
