% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vardetail.R
\name{vardetails}
\alias{vardetails}
\title{Check vars details}
\usage{
vardetails(id)
}
\arguments{
\item{id}{Variable's ID. It can be a single string or a vector.}
}
\value{
A tibble
}
\description{
Check the details of a variable, by searching for it's id.
}
\examples{
vardetails(id = 4049)
vardetails(id = c(4049,4031))

}
