% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vars.R
\name{vars}
\alias{vars}
\title{List, filter and sort variables' ID and name}
\usage{
vars(contains = NULL, sort = "ASC")
}
\arguments{
\item{contains}{Optional. A character string to filter from the list of variables.}

\item{sort}{Optional. A character string. Either "ASC" to sort from A to Z or "DESC" to sort from Z to A.}
}
\value{
a tibble
}
\description{
Lists, filters and sorts variables' ID and names.
}
\examples{
#Run this to see the full list of variables
vars()

#You can also add arguments to filter and sort the list of variables
vars(sort = "ASC",contains = "Agricultura")




}
